+++ src/main/java/br/gov/ifgoiano/gethospeda/service/QuartoService.java
@@ -1,8 +1,11 @@
 package br.gov.ifgoiano.gethospeda.service;
 
+import br.gov.ifgoiano.gethospeda.controller.ImovelController;
+import br.gov.ifgoiano.gethospeda.controller.QuartoController;
 import br.gov.ifgoiano.gethospeda.dto.QuartoCompletoDTO;
 import br.gov.ifgoiano.gethospeda.dto.QuartoCreateDTO;
 import br.gov.ifgoiano.gethospeda.dto.QuartoResumoDTO;
+import br.gov.ifgoiano.gethospeda.dto.QuartoUpdateDTO;
 import br.gov.ifgoiano.gethospeda.exception.ResourceNotFoundException;
 import br.gov.ifgoiano.gethospeda.model.Quarto;
 import br.gov.ifgoiano.gethospeda.repository.QuartoRepository;
@@ -14,6 +17,9 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.logging.Logger;
 
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
+
 @Service
 public class QuartoService {
     @Autowired
@@ -23,7 +29,11 @@ public class QuartoService {
     public List<QuartoResumoDTO> findAll() {
         logger.info("findAll");
         List<Quarto> quartos = quartoRepository.findAll();
-        return DataMapper.parseListObjects(quartos, QuartoResumoDTO.class);
+        List<QuartoResumoDTO> quartoResumoDTOS = DataMapper.parseListObjects(quartos, QuartoResumoDTO.class);
+        quartoResumoDTOS.forEach(quartoResumoDTO -> {
+            quartoResumoDTO.add(linkTo(methodOn(QuartoController.class).findById(quartoResumoDTO.getId())).withSelfRel());
+        });
+        return quartoResumoDTOS;
     }
 
     public List<QuartoResumoDTO> findByImovel(long id) {
@@ -36,7 +46,9 @@ public class QuartoService {
     public QuartoCompletoDTO findById(Long id) {
         logger.info("findById");
         Quarto quarto = quartoRepository.findById(id).orElseThrow(()-> new ResourceNotFoundException("No records found for this ID"));
-        return DataMapper.parseObject(quarto, QuartoCompletoDTO.class);
+        QuartoCompletoDTO quartoCompletoDTO =  DataMapper.parseObject(quarto, QuartoCompletoDTO.class);
+        quartoCompletoDTO.getImovel().add(linkTo(methodOn(ImovelController.class).findById(quartoCompletoDTO.getImovel().getId())).withSelfRel());
+        return quartoCompletoDTO;
     }
 
     public QuartoCompletoDTO save(QuartoCreateDTO quarto) {
@@ -44,7 +56,7 @@ public class QuartoService {
         return DataMapper.parseObject(quartoRepository.save(DataMapper.parseObject(quarto,Quarto.class)),QuartoCompletoDTO.class);
     }
 
-    public QuartoCompletoDTO update(QuartoCompletoDTO quarto) {
+    public QuartoCompletoDTO update(QuartoUpdateDTO quarto) {
         logger.info("update");
         Quarto newQuarto = quartoRepository.findById(quarto.getId()).orElseThrow(()-> new ResourceNotFoundException("No records found for this ID"));
         newQuarto.setTipo(quarto.getTipo());
@@ -54,8 +66,13 @@ public class QuartoService {
         newQuarto.setDescricao(quarto.getDescricao());
         return  DataMapper.parseObject(quartoRepository.save(newQuarto),QuartoCompletoDTO.class);
     }
-    public void deleteById(Long id) {
+    public boolean deleteById(Long id) {
         logger.info("deleteById");
+        if (quartoRepository.findById(id).isPresent()) {
         quartoRepository.deleteById(id);
+        return true;
+        }else {
+            return false;
+        }
     }
 }
